+++ include/coli/game-engine.h
@@ -1,6 +1,7 @@
 #ifndef COLI_GAME_ENGINE_H
 #define COLI_GAME_ENGINE_H
 
+#include "coli/version.h"
 #include "coli/utility.h"
 
 #include "coli/game/components/layer.h"
@@ -9,5 +10,8 @@
 
 #include "coli/generic/system.h"
 #include "coli/generic/engine.h"
+
+#include "coli/graphics/context.h"
+#include "coli/graphics/window.h"
 
 #endif