+++ CHANGES.md
@@ -9,7 +9,18 @@
 
 ## Changelogs
 
-### v0.1.3 - **Current**
+### v0.2.0 - **Current**
+
+- Graphics:
+  - New base class for graphic resources 
+  - Added base for graphics system (context and window) powered by OpenGL
+  - New dependency [`glfw3`](https://github.com/glfw/glfw)
+  - New target `coli-glad`. The source code was generated [here](https://glad.dav1d.de/)
+- Tests:
+  - Added context test
+  - Also added a few window tests
+
+### v0.1.3
 
 - QoL:
   - The library version is provided to code via dynamically generated `version.h`